refactor(api): 从 LongTermMemory 拆分 MemoryConsumer 至独立模块
将 _consumer 方法体、6 个 @tool 函数、模块级 _current_mm 状态、
以及冷启动/更新提示词常量迁至新的 api/memory/consumer.py。

LongTermMemory 从 675 行削至 352 行，MemoryConsumer 负责：
- 后台消费管线：WebSocket 通知 → CRUD Agent → memory.yaml 写入
- CRUD 工具的模块级状态和生命周期
